@Silicon27 2025-03-13
My benchmarks show: 

Old:
```
Command Benchmark Statistics
---------------------------
Iterations: 100
Total Time: 2.261 s
Average ± Std Dev: 22.608 ms ± 2.451 ms
Minimum Time: 21.346 ms
Maximum Time: 46.409 ms
```

with your changes:
```
Command Benchmark Statistics
---------------------------
Iterations: 100
Total Time: 2.268 s
Average ± Std Dev: 22.680 ms ± 3.963 ms
Minimum Time: 21.132 ms
Maximum Time: 60.899 ms
```

Hmm, yk, it's not a matter of performance now; it's more about cache hits and how the files on your system are named/structured. As far as Python is concerned, performance is definitely not something you can achieve with vanilla. I'll get a C++ out ASAP.